Web3Impl refactor, extrating classes
