+++ rskj-core/src/main/java/org/ethereum/rpc/Web3Impl.java
@@ -942,153 +942,6 @@ public class Web3Impl implements Web3 {
         throw new UnsupportedOperationException("Serpent compiler not supported");
     }
 
-    static class Filter {
-        abstract static class FilterEvent {
-            public abstract Object getJsonEventObject();
-        }
-
-        List<FilterEvent> events = new ArrayList<>();
-        int processedEvents = 0;
-        long accessTime = System.currentTimeMillis();
-
-        public boolean hasExpired(long timeout) {
-            long nowTime = System.currentTimeMillis();
-
-            return accessTime + timeout <= nowTime;
-        }
-
-        public synchronized Object[] getNewEvents() {
-            this.accessTime = System.currentTimeMillis();
-
-            Object[] ret = events.stream().skip(processedEvents).map(fe -> fe.getJsonEventObject()).collect(Collectors.toList()).toArray();
-
-            processedEvents = events.size();
-
-            return ret;
-        }
-
-        public synchronized Object[] getEvents() {
-            this.accessTime = System.currentTimeMillis();
-
-            return events.stream().map(fe -> fe.getJsonEventObject()).collect(Collectors.toList()).toArray();
-        }
-
-        protected synchronized void add(FilterEvent evt) {
-            events.add(evt);
-        }
-
-        public void newBlockReceived(Block b) {
-        }
-
-        public void newPendingTx(Transaction tx) {
-            // add TransactionReceipt for PendingTx
-        }
-    }
-
-    static class NewBlockFilter extends Filter {
-        class NewBlockFilterEvent extends FilterEvent {
-            public final Block b;
-
-            NewBlockFilterEvent(Block b) {
-                this.b = b;
-            }
-
-            @Override
-            public String getJsonEventObject() {
-                return toJsonHex(b.getHash());
-            }
-        }
-
-        @Override
-        public void newBlockReceived(Block b) {
-            add(new NewBlockFilterEvent(b));
-        }
-    }
-
-    static class PendingTransactionFilter extends Filter {
-        class PendingTransactionFilterEvent extends FilterEvent {
-            private final Transaction tx;
-
-            PendingTransactionFilterEvent(Transaction tx) {
-                this.tx = tx;
-            }
-
-            @Override
-            public String getJsonEventObject() {
-                return toJsonHex(tx.getHash());
-            }
-        }
-
-        @Override
-        public void newPendingTx(Transaction tx) {
-            add(new PendingTransactionFilterEvent(tx));
-        }
-    }
-
-    class JsonLogFilter extends Filter {
-        class LogFilterEvent extends FilterEvent {
-            private final LogFilterElement el;
-
-            LogFilterEvent(LogFilterElement el) {
-                this.el = el;
-            }
-
-            @Override
-            public LogFilterElement getJsonEventObject() {
-                return el;
-            }
-        }
-
-        LogFilter logFilter;
-        boolean onNewBlock;
-        boolean onPendingTx;
-
-        public JsonLogFilter(LogFilter logFilter) {
-            this.logFilter = logFilter;
-        }
-
-        void onLogMatch(LogInfo logInfo, Block b, int txIndex, Transaction tx, int logIdx) {
-            add(new LogFilterEvent(new LogFilterElement(logInfo, b, txIndex, tx, logIdx)));
-        }
-
-        void onTransaction(Transaction tx, Block b, int txIndex) {
-            TransactionInfo txInfo = blockchain.getTransactionInfo(tx.getHash());
-            TransactionReceipt receipt = txInfo.getReceipt();
-
-            LogFilterElement[] logs = new LogFilterElement[receipt.getLogInfoList().size()];
-
-            for (int i = 0; i < logs.length; i++) {
-                LogInfo logInfo = receipt.getLogInfoList().get(i);
-                if (logFilter.matchesContractAddress(logInfo.getAddress())) {
-                    onLogMatch(logInfo, b, txIndex, receipt.getTransaction(), i);
-                }
-            }
-        }
-
-        void onBlock(Block b) {
-            if (logFilter.matchBloom(new Bloom(b.getLogBloom()))) {
-                int txIdx = 0;
-
-                for (Transaction tx : b.getTransactionsList()) {
-                    onTransaction(tx, b, txIdx);
-                    txIdx++;
-                }
-            }
-        }
-
-        @Override
-        public void newBlockReceived(Block b) {
-            if (onNewBlock) {
-                onBlock(b);
-            }
-        }
-
-        @Override
-        public void newPendingTx(Transaction tx) {
-            //empty method
-        }
-    }
-
     AtomicInteger filterCounter = new AtomicInteger(1);
     Map<Integer, Filter> installedFilters = new Hashtable<>();
 
@@ -1134,7 +987,7 @@ public class Web3Impl implements Web3 {
                 }
             }
 
-            JsonLogFilter filter = new JsonLogFilter(logFilter);
+            JsonLogFilter filter = new JsonLogFilter(logFilter, blockchain);
 
             int id;
